Created the CalculatorResults view and Controller. (Not in use yet) Modified the Armor model to work with the CalculatorDisplayController more efficiently.
The "calculate" button on the CalculatorDisplay view now functions and will print out 3 values to the console.
It will print the total protection of the helmet and chestplate pieces.
It will print the total protection of the leggings and boots pieces.
It will then add the two values together to finally print the total armor protection.

Will be revising how the armor material choices get their protection value.
WIll continue working on implementing the CalculatorResults view which gets its values from the "calculate" button.